Reorganize navigation: Settings breakout and Library consolidation
- Break out Settings tab from ProfileScreen into standalone SettingsScreen
- Add UserSettingsViewModel for settings functionality
- Move Collections and Reading List access into Library categories
- Remove Collections and Reading List from user dropdown menu
- ProfileScreen now shows only profile content (stats, top authors, genres)
- SettingsScreen contains Playback, Account, About sections and Logout
